Long running thread tests
